% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tarmander.R
\name{tarmander}
\alias{tarmander}
\title{A microutil for switching between \code{targets} scripts}
\usage{
tarmander()
}
\description{
You can call the function object to print the current active project or call
the actual function to select/change projects.
}
\examples{
tarmander::tarmander()
}
